+++ packages/plugins/AddressSearch/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Feature: There is now a new optional configuration parameter `searchMethods.queryParameters.likeFilterAttributes` for the WfsSearch that allows to add custom key/value pairs which are mapped to attributes of the like filter operator.
 - Feature: `AddressSearchState`, `AddressSearchGetters`, and `MpApiParameters` types have been additionally exposed as root export since using packages frequently rely on them.
 - Feature: When using the search type `mpapi`, a second search using a wildcard is now being triggered when the first search yielded no results.
+- Feature: Custom select functions are now called with the store's `this` context, allowing for more flexible data access.
 - Fix: Adjust documentation and types to properly describe optionality of configuration parameters.
 - Fix: Use correct getter for `minLength`.
 - Fix: `SearchResultSymbols` has been additionally exposed as root export since using packages frequently rely on it.